Adding method to obtain IMC

+++ src/main/java/com/gym/gym_app/dto/UsuarioResponse.java
@@ -3,13 +3,13 @@ package com.gym.gym_app.dto;
 import com.gym.gym_app.models.Rol;
 
 public class UsuarioResponse {
-    private Integer id;
+    private Long id;
     private String nombre;
     private String correo;
     private String telefono;
     private Rol rol;
 
-    public UsuarioResponse(Integer id, String nombre, String correo, String telefono, Rol rol) {
+    public UsuarioResponse(Long id, String nombre, String correo, String telefono, Rol rol) {
         this.id = id;
         this.nombre = nombre;
         this.correo = correo;
@@ -17,7 +17,7 @@ public class UsuarioResponse {
         this.rol = rol;
     }
 
-    public Integer getId() {
+    public Long getId() {
         return id;
     }
